// src/types/slices/complaintChannel.ts
export interface ComplaintChannel {
    id: number
    identifier: number
    name: string
    description: string
    iduser: number
    user_name: string
    n_document_user: string
    idcompany: number
    company: string
    registry_number_company: string
    status: string
    created_at: string
}